requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
sqlalchemy==2.0.44
alembic==1.17.0
psycopg2-binary==2.9.11
redis==7.0.0
python-dotenv==1.1.1
pydantic[email]==2.9.0
email-validator==2.1.1
